requirements.txt
apache-airflow==2.5.0
requests
psycopg2-binary
